+++ packages/gatsby-source-wordpress-experimental/src/models/gatsby-api.js
@@ -25,7 +25,7 @@ const defaultPluginOptions = {
   },
   schema: {
     queryDepth: 15,
-    querySelfAncestorLimit: 3,
+    circularQueryLimit: 3,
     typePrefix: `Wp`,
     timeout: 30 * 1000, // 30 seconds
   },